+++ lib/generators/dbviewer/install_generator.rb
@@ -1,6 +1,6 @@
 module Dbviewer
   module Generators
-    class InitializerGenerator < Rails::Generators::Base
+    class InstallGenerator < Rails::Generators::Base
       source_root File.expand_path("../templates", __FILE__)
       desc "Creates a DBViewer initializer file at config/initializers/dbviewer.rb"